add authentication to delete request

+++ src/main/java/org/marketplace/services/AdvertisementManagementService.java
@@ -3,7 +3,9 @@ package org.marketplace.services;
 import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityNotFoundException;
 import org.marketplace.models.Advertisement;
+import org.marketplace.models.User;
 import org.marketplace.repositories.AdvertisementManagementRepository;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,15 +15,18 @@ import java.util.Optional;
 public class AdvertisementManagementService {
     private final AdvertisementManagementRepository advertisementManagementRepository;
 
-    public AdvertisementManagementService(AdvertisementManagementRepository advertisementManagementRepository) {
+    private final AuthorizationUserUtil authorizationUserUtil;
+
+    public AdvertisementManagementService(AdvertisementManagementRepository advertisementManagementRepository, AuthorizationUserUtil authorizationUserUtil) {
         this.advertisementManagementRepository = advertisementManagementRepository;
+        this.authorizationUserUtil = authorizationUserUtil;
     }
 
     public Advertisement addAdvertisement(Advertisement advertisement) {
         try {
             getAdvertisementById(advertisement.getId());
             throw new EntityExistsException(String.format("Advertisement with id: %d already exsits!", advertisement.getId()));
-        } catch(EntityNotFoundException e) {
+        } catch (EntityNotFoundException e) {
             return advertisementManagementRepository.save(advertisement);
         }
     }
@@ -30,16 +35,18 @@ public class AdvertisementManagementService {
         try {
             getAdvertisementById(advertisement.getId());
             return advertisementManagementRepository.save(advertisement);
-        } catch(EntityNotFoundException e) {
+        } catch (EntityNotFoundException e) {
             throw new EntityNotFoundException(String.format("Advertisement with id: %d was not found", advertisement.getId()));
         }
     }
 
-    public void deleteAdvertisement(Long id) {
-        if(!advertisementManagementRepository.existsById(id)){
+    public void deleteAdvertisement(Long id, String token) {
+        if (!advertisementManagementRepository.existsById(id)) {
             throw new EntityNotFoundException(String.format("Advertisement with id: %d was not found", id));
         }
-
+        User user = getAdvertisementById(id).getUser();
+        if (!authorizationUserUtil.checkAccessToUserByCurrentUser(token, user))
+            throw new AccessDeniedException(String.format("You are not authorized to delete advertisement with id: ", id));
         advertisementManagementRepository.deleteById(id);
     }
 
@@ -49,7 +56,7 @@ public class AdvertisementManagementService {
 
     public Advertisement getAdvertisementById(Long id) {
         Optional<Advertisement> advertisement = advertisementManagementRepository.findById(id);
-        if(advertisement.isEmpty()){
+        if (advertisement.isEmpty()) {
             throw new EntityNotFoundException(String.format("Advertisement with id: %d was not found", id));
         }
